requirements.txt
streamlit>=1.32.0
groq>=0.9.0
langchain-groq>=0.1.3
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
gTTS>=2.5.0
